Restructure source to common preact layout
